--- src/main/java/com/spotify/confidence/FlushPolicy.java
@@ -1,9 +1,0 @@
-package com.spotify.confidence;
-
-interface FlushPolicy {
-  void hit();
-
-  boolean shouldFlush();
-
-  void reset();
-}
