+++ src/main/java/com/spotify/confidence/eventsender/ConfidenceValue.java
@@ -1,4 +1,4 @@
-package com.spotify.confidence;
+package com.spotify.confidence.eventsender;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
