+++ src/main/java/com/spotify/confidence/eventsender/Event.java
@@ -1,0 +1,13 @@
+package com.spotify.confidence.eventsender;
+
+class Event {
+  private final String name;
+  private final ConfidenceValue.Struct message;
+  private final ConfidenceValue.Struct context;
+
+  public Event(String name, ConfidenceValue.Struct message, ConfidenceValue.Struct context) {
+    this.name = name;
+    this.message = message;
+    this.context = context;
+  }
+}
