+++ src/main/java/com/spotify/confidence/eventsender/EventSenderEngine.java
@@ -1,4 +1,5 @@
-package com.spotify.confidence;
+package com.spotify.confidence.eventsender;
+
 
 interface EventSenderEngine {
   void send(String name, ConfidenceValue.Struct message, ConfidenceValue.Struct context);
