+++ src/main/java/com/spotify/confidence/eventsender/EventSenderEngineImpl.java
@@ -1,64 +1,13 @@
-package com.spotify.confidence;
+package com.spotify.confidence.eventsender;
 
 import com.google.common.collect.ImmutableMap;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 import java.util.concurrent.*;
 
 class EventSenderEngineImpl implements EventSenderEngine {
   private final BlockingQueue<Event> writeQueue = new LinkedBlockingQueue<>();
   private final BlockingQueue<String> uploadQueue = new LinkedBlockingQueue<>();
-  private final EventSenderStorage eventStorage =
-      new EventSenderStorage() {
-        private final List<Event> events = new ArrayList<>();
-        private final List<EventBatch> readyBatches = new ArrayList<>();
-        final Semaphore semaphore = new Semaphore(1);
-
-        @Override
-        public void write(Event event) {
-          runWithSemaphore(
-              () -> {
-                this.events.add(event);
-              });
-        }
-
-        void runWithSemaphore(Runnable codeBlock) {
-          try {
-            semaphore.acquire(); // Acquire the semaphore, blocking if necessary
-            // Execute the provided code block
-            codeBlock.run();
-          } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-          } finally {
-            semaphore.release(); // Release the semaphore after executing the code block
-          }
-        }
-
-        @Override
-        public void batch() {
-          runWithSemaphore(
-              () -> {
-                EventBatch batch = new EventBatch(List.copyOf(events));
-                events.clear();
-                readyBatches.add(batch);
-              });
-        }
-
-        @Override
-        public List<EventBatch> readyEvents() {
-          return readyBatches;
-        }
-
-        @Override
-        public void deleteBatch(String batchId) {
-          runWithSemaphore(
-              () -> {
-                readyBatches.removeIf(eventBatch -> eventBatch.id().equals(batchId));
-                System.out.println("deleted");
-              });
-        }
-      };
+  private final EventSenderStorage eventStorage = new InMemoryStorage();
   private final List<FlushPolicy> flushPolicies;
   private boolean isStopped = false;
 
@@ -145,43 +94,4 @@ class EventSenderEngineImpl implements EventSenderEngine {
   public void send(String name, ConfidenceValue.Struct context) {
     send(name, ConfidenceValue.of(ImmutableMap.of()), context);
   }
-}
-
-class Event {
-  private final String name;
-  private final ConfidenceValue.Struct message;
-  private final ConfidenceValue.Struct context;
-
-  public Event(String name, ConfidenceValue.Struct message, ConfidenceValue.Struct context) {
-    this.name = name;
-    this.message = message;
-    this.context = context;
-  }
-}
-
-class EventBatch {
-  private final List<Event> events;
-  private String id = UUID.randomUUID().toString();
-
-  public EventBatch(List<Event> events) {
-    this.events = events;
-  }
-
-  public String id() {
-    return id;
-  }
-
-  public List<Event> events() {
-    return events;
-  }
-}
-
-interface EventSenderStorage {
-  void write(Event event);
-
-  void batch();
-
-  List<EventBatch> readyEvents();
-
-  void deleteBatch(String batchId);
 }
