+++ src/main/java/com/spotify/confidence/eventsender/EventSenderStorage.java
@@ -1,0 +1,13 @@
+package com.spotify.confidence.eventsender;
+
+import java.util.List;
+
+interface EventSenderStorage {
+  void write(Event event);
+
+  void batch();
+
+  List<EventBatch> readyEvents();
+
+  void deleteBatch(String batchId);
+}
