+++ src/main/java/com/spotify/confidence/eventsender/EventUploader.java
@@ -1,0 +1,7 @@
+package com.spotify.confidence.eventsender;
+
+import java.util.concurrent.CompletableFuture;
+
+interface EventUploader {
+  CompletableFuture<Boolean> upload(EventBatch event);
+}
