+++ src/main/java/com/spotify/confidence/eventsender/FlushPolicy.java
@@ -1,0 +1,9 @@
+package com.spotify.confidence.eventsender;
+
+public interface FlushPolicy {
+  void hit();
+
+  boolean shouldFlush();
+
+  void reset();
+}
